@@ -1,45 +1,49 @@
 <?php
 
-namespace App\Http\Controllers;
-/** @var \Laravel\Lumen\Routing\Router $router */
+use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\OrderController;
+use App\Http\Controllers\ProductController;
 
 /*
 |--------------------------------------------------------------------------
-| Application Routes
+| Web Routes
 |--------------------------------------------------------------------------
 |
-| Here is where you can register all of the routes for an application.
-| It is a breeze. Simply tell Lumen the URIs it should respond to
-| and give it the Closure to call when that URI is requested.
+| Here is where you can register web routes for your application. These
+| routes are loaded by the RouteServiceProvider within a group which
+| contains the "web" middleware group. Now create something great!
 |
 */
 
-$router->get('/', function () use ($router) {
-    return $router->app->version();
+Route::get('/', function () {
+    return view('welcome');
 });
 
-    $router->group(['prefix' => 'api'], function () use ($router) {
-        //users
-        $router->get('users', 'UserController@showUsers');
-        $router->post('createUser', 'UserController@createUser');
-        $router->get('editUser/{id}', 'UserController@editUser');
-        $router->put('updateUser/{id}', 'UserController@updateUser');
-        $router->delete('deleteUser/{id}', 'UserController@deleteUser');
 
-        //products
+    //users
+    Route::prefix('api')->group(function () {
 
-        $router->get('products', 'ProductController@showProducts');
-        $router->post('createProduct', 'ProductController@createProduct');
-        $router->get('editProduct/{id}', 'ProductController@editProduct');
-        $router->put('updateProduct/{id}', 'ProductController@updateProduct');
-        $router->delete('deleteProduct/{id}', 'ProductController@deleteProduct');
+    ROUTE::get('users', [UserController::class, 'showUsers']);
+    ROUTE::post('createUser', [UserController::class, 'createUser']);
+    ROUTE::get('editUser/{id}', [UserController::class, 'editUser']);
+    ROUTE::put('updateUser/{id}', [UserController::class, 'updateUser']);
+    ROUTE::delete('deleteUser/{id}', [UserController::class, 'deleteUser']);
 
-        //orders
+    
+    //products
 
-        $router->get('orders', 'OrderController@showOrders');
-        $router->post('createOrder', 'OrderController@createOrder');
-        $router->get('editOrder/{id}', 'OrderController@editOrder');
-        $router->put('updateOrder/{id}', 'OrderController@updateOrder');
-        $router->delete('deleteOrder/{id}', 'OrderController@deleteOrder');
-    });
+    ROUTE::get('products', [ProductController::class, 'showProducts']);
+    ROUTE::post('createProduct', [ProductController::class, 'createProduct']);
+    ROUTE::get('editProduct/{id}', [ProductController::class, 'editProduct']);
+    ROUTE::put('updateProduct/{id}', [ProductController::class, 'updateProduct']);
+    ROUTE::delete('deleteProduct/{id}', [ProductController::class, 'deleteProduct']);
 
+    //orders
+
+    ROUTE::get('orders', [OrderController::class, 'showOrders']);
+    ROUTE::post('createOrder', [OrderController::class, 'createOrder']);
+    ROUTE::get('editOrder/{id}', [OrderController::class, 'editOrder']);
+    ROUTE::put('updateOrder/{id}', [OrderController::class, 'updateOrder']);
+    ROUTE::delete('deleteOrder/{id}', [OrderController::class, 'deleteOrder']);
+});
