@@ -2,12 +2,13 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use App\Models\Order;
-use App\Models\Image;
 
 class Product extends Model
 {
+    use HasFactory;
+
     protected $table = 'products';
     protected $primaryKey = 'id';
     protected $fillable = [
