@@ -2,12 +2,13 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use App\Models\City;
-use App\Models\Region;
 
 class Province extends Model
 {
+    use HasFactory;
+
     protected $table = 'provinces';
     protected $primaryKey = 'id';
     protected $fillable = [
