@@ -1,12 +1,14 @@
 <?php
 
 namespace App\Models;
-use App\Models\Province;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Region extends Model
 {
+    use HasFactory;
+
     protected $table = 'region';
     protected $primaryKey = 'id';
     protected $fillable = [
